requirements.txt
discord.py==1.4.1
toml==0.10.1